fix(ci): consolidate CI fixes into a single commit

+++ backup.sh
@@ -2,23 +2,24 @@
 
 . ./includes.sh
 
-function backup_init() {
-    NOW="$(date +"${BACKUP_FILE_DATE_FORMAT}")"
-    # backup vaultwarden database file
-    BACKUP_FILE_DB="${BACKUP_DIR}/db.sqlite3"
-    # backup vaultwarden config file
-    BACKUP_FILE_CONFIG="${BACKUP_DIR}/config.json"
-    # backup vaultwarden rsakey files
-    BACKUP_FILE_RSAKEY="${BACKUP_DIR}/rsakey.tar"
-    # backup vaultwarden attachments directory
-    BACKUP_FILE_ATTACHMENTS="${BACKUP_DIR}/attachments.tar"
-    # backup vaultwarden sends directory
-    BACKUP_FILE_SENDS="${BACKUP_DIR}/sends.tar"
-    # backup temp file
-    BACKUP_FILE_TEMP="${BACKUP_DIR}/backup.temp.tar"
-    # backup gzip file
-    BACKUP_FILE_GZIP="${BACKUP_DIR}/backup.${NOW}.tar.gz"
-}
+
+NOW="$(date +"${BACKUP_FILE_DATE_FORMAT}")"
+# backup vaultwarden database file
+BACKUP_FILE_DB="${BACKUP_DIR}/db.sqlite3"
+# backup vaultwarden config file
+BACKUP_FILE_CONFIG="${BACKUP_DIR}/config.json"
+# backup vaultwarden rsakey files
+BACKUP_FILE_RSAKEY="${BACKUP_DIR}/rsakey.tar"
+# backup vaultwarden attachments directory
+BACKUP_FILE_ATTACHMENTS="${BACKUP_DIR}/attachments.tar"
+# backup vaultwarden sends directory
+BACKUP_FILE_SENDS="${BACKUP_DIR}/sends.tar"
+# backup temp file
+BACKUP_FILE_TEMP="${BACKUP_DIR}/backup.temp.tar"
+# backup gzip file
+BACKUP_FILE_GZIP="${BACKUP_DIR}/backup.${NOW}.tar.gz"
+# Getting public key from the age key file
+PUBLIC_KEY="$(cat "${AGE_KEY_FILE}" | awk 'match($0, /^# *public key: */) {gsub(/^# *public key: */, ""); print}')"
 
 function backup_db() {
     echo "backup vaultwarden sqlite database"
@@ -44,8 +45,10 @@ function backup_rsakey() {
     echo "backup vaultwarden rsakey ${DATA_RSAKEY_DIRNAME}"
 
     if [[ -d "${DATA_RSAKEY_DIRNAME}" ]]; then
-        local FIND_RSAKEY=$(find "${DATA_RSAKEY_DIRNAME}" -name "${DATA_RSAKEY_BASENAME}*" | xargs -I {} basename {})
-        local FIND_RSAKEY_COUNT=$(echo "${FIND_RSAKEY}" | wc -l)
+        local FIND_RSAKEY
+        FIND_RSAKEY=$(find "${DATA_RSAKEY_DIRNAME}" -name "${DATA_RSAKEY_BASENAME}*" -print0 | xargs -0 -I {} basename {})
+        local FIND_RSAKEY_COUNT
+        FIND_RSAKEY_COUNT=$(echo "${FIND_RSAKEY}" | wc -l)
 
         if [[ "${FIND_RSAKEY_COUNT}" -gt 0 ]]; then
             echo "${FIND_RSAKEY}" | tar -c -C "${DATA_RSAKEY_DIRNAME}" -f "${BACKUP_FILE_RSAKEY}" -T -
@@ -82,12 +85,10 @@ function backup_sends() {
 
 function backup_package() {
 
-    echo "package backup file"
+    echo "package backup file and encryption with age"
 
-    UPLOAD_FILE="${BACKUP_FILE_GZIP}"
-    tar -cf "${BACKUP_FILE_TEMP}" "${BACKUP_DIR}"/*
-    gzip "${BACKUP_FILE_TEMP}"
-    mv "${BACKUP_FILE_TEMP}.gz" "${BACKUP_FILE_GZIP}"
+
+    tar -czf - "${BACKUP_DIR}"/*  | age -e -r "$PUBLIC_KEY" > "$BACKUP_FILE_GZIP"
 
 }
 
@@ -106,21 +107,20 @@ function upload() {
 
     #update lifecycle configuration
     envsubst < lifecycle.json.tmpl > lifecycle.json
-    aws s3api put-bucket-lifecycle-configuration --bucket $S3_BUCKET --endpoint-url $S3_ENDPOINT --lifecycle-configuration file://lifecycle.json
+    aws s3api put-bucket-lifecycle-configuration --bucket $S3_BUCKET --lifecycle-configuration file://lifecycle.json
 
     echo "Uploading dump to bucket $S3_BUCKET"
-    aws s3 $AWS_ARGS cp ${BACKUP_FILE_GZIP} s3://$S3_BUCKET/$S3_PREFIX/ || exit 2
+    aws s3 cp "${BACKUP_FILE_GZIP}" s3://"$S3_BUCKET"/"$S3_PREFIX"/ || exit 2
 
 }
 
 function cleanup() {
-    /bin/rm -rf ${BACKUP_DIR}/*
+    /bin/rm -rf "${BACKUP_DIR:?}"/*
 }
 
 
 init_env
 cleanup
-backup_init
 backup
 backup_package
 upload
